moved comments into different mock type
