+++ src/mocks/film-card.js
@@ -5,7 +5,7 @@ import {
   getRandomArrayItem,
   getRandomBooleanValue,
   generateDate} from '../helpers/common';
-
+import {generateComments} from "./comment";
 const FilmTitles = [
   `Apocalypse Now`,
   `Blade Runner`,
@@ -42,35 +42,7 @@ const Genres = [
   `Western`,
 ];
 
-const Emotions = [
-  `angry`,
-  `puke`,
-  `sleeping`,
-  `smile`,
-];
-
-const Users = [
-  `Vasya`,
-  `Druzhok-Kuraek`,
-  `Armin Van Buuren`,
-  `Shia LaBeouf`,
-  `Christopher Lawrence`,
-  `Laure Sainclair`,
-  `Lizzy Burden`,
-  `Ostap Bender`,
-  `Johan Gielen`,
-  `Markus Schulz`,
-  `Joyful Grape`,
-  `Jason Statham`,
-];
 const generateId = () => Math.random().toString(36).substr(2, 9);
-
-const getRandomCommentDate = () => {
-  const currentDate = Date.now();
-  const threeDaysInMs = 1000 * 60 * 60 * 24 * 3;
-  const diffDate = getRandomIntInclusive(0, threeDaysInMs);
-  return new Date(currentDate - diffDate);
-};
 
 const generateDescription = () => {
   const sentences = text
@@ -92,25 +64,6 @@ const generateDescription = () => {
 const generateGenres = (genres) => genres
   .filter(getRandomBooleanValue)
   .slice(0, getRandomIntInclusive(1, 3));
-
-const generateComment = () => {
-  return {
-    text: generateDescription(),
-    emotions: getRandomArrayItem(Emotions),
-    author: getRandomArrayItem(Users),
-    date: getRandomCommentDate(),
-  };
-};
-
-const generateComments = () => {
-  const commentsAmount = getRandomIntInclusive(0, 5);
-  const result = [];
-  for (let i = 0; i < commentsAmount; i++) {
-    result.push(generateComment());
-  }
-
-  return result;
-};
 
 const generateFilmCard = () => {
   const rating = getRandomErratic(1, 10);
@@ -141,4 +94,4 @@ const generateFilmCards = (count) => {
   return result;
 };
 
-export {generateFilmCard, generateFilmCards};
+export {generateFilmCard, generateFilmCards, generateDescription};
